mitm-injector/requirements.txt
mitmproxy>=10.0
PyYAML>=6.0